Add some casts-to-int to keep the compiler quiet.
(bzr r2369)

+++ src/libnr/nr-rect-l.cpp
@@ -2,8 +2,8 @@
 #include <math.h>
 
 namespace NR{
-    IRect::IRect(const Rect& r) : _min(floor(r.min()[X]), floor(r.min()[Y])), 
-				  _max(ceil(r.min()[X]), ceil(r.min()[Y])) 
+    IRect::IRect(const Rect& r) : _min(int(floor(r.min()[X])), int(floor(r.min()[Y]))),
+				  _max(int(ceil(r.min()[X])), int(ceil(r.min()[Y])))
     {
 	    
     }
